# improve sync timeouts by being more conservative the fewer peers we have